Fix line execution strategy in a block.
Do not automatically execute new blocks when encountered,
as these blocks might be if/else blocks, or while true blocks.
The next line decision *MUST* come from Lexicon (who will always return
the right next line to execute, even in the case of a new block within
the current block). We never execute a line outside of our block because
of the line_has_parent check in the while loop.

This was done to fix the following behaviour:
When a when block had an if/else, it would execute both branches of it
because the condition was never evaluated (because it just checked for
an enter line number).