Pin versions of the test deps to that mentioned in the tox.ini.
The motivation for this is to enable users to simply clone the project, and then be able to run python setup.py test. Will and I explored another option of forcing users to run tox instead of the setuptools test, but were unable to successfully override the test command and print a decent message. Hence, for now, we'll just pin the same versions (since they rarely change anyway).