Iterate over all lines using the compilation tree directly.
App#start_services needs to iterate through every single line to find out services which it needs to start beforehand (at the time of deployment). However, with Storyscript compiler 0.25.x, next no longer indicates the actual next line, but it respects the scope. With regard to this, no services are discovered to be bootstrapped.